# perf(NODE-6356): Improve serialization performance