--- a/src/constants.ts
+++ b/src/constants.ts
@@ -2,6 +2,9 @@
 export const BSON_MAJOR_VERSION = 6;
 
 /** @internal */
+export const BSON_VERSION_SYMBOL = Symbol.for('@@mdb.bson.version');
+
+/** @internal */
 export const BSON_INT32_MAX = 0x7fffffff;
 /** @internal */
 export const BSON_INT32_MIN = -0x80000000;