--- a/src/extended_json.ts
+++ b/src/extended_json.ts
@@ -6,7 +6,8 @@ import {
   BSON_INT32_MIN,
   BSON_INT64_MAX,
   BSON_INT64_MIN,
-  BSON_MAJOR_VERSION
+  BSON_MAJOR_VERSION,
+  BSON_VERSION_SYMBOL
 } from './constants';
 import { DBRef, isDBRefLike } from './db_ref';
 import { Decimal128 } from './decimal128';
@@ -358,7 +359,7 @@ function serializeDocument(doc: any, options: EJSONSerializeOptions) {
     doc != null &&
     typeof doc === 'object' &&
     typeof doc._bsontype === 'string' &&
-    doc[Symbol.for('@@mdb.bson.version')] !== BSON_MAJOR_VERSION
+    doc[BSON_VERSION_SYMBOL] !== BSON_MAJOR_VERSION
   ) {
     throw new BSONVersionError();
   } else if (isBSONType(doc)) {