--- a/src/parser/utils.ts
+++ b/src/parser/utils.ts
@@ -1,5 +1,16 @@
 const map = new WeakMap<object, string>();
 
+const TYPES = {
+  ArrayBuffer: '[object ArrayBuffer]',
+  SharedArrayBuffer: '[object SharedArrayBuffer]',
+  Uint8Array: '[object Uint8Array]',
+  BigInt64Array: '[object BigInt64Array]',
+  BigUint64Array: '[object BigUint64Array]',
+  RegExp: '[object RegExp]',
+  Map: '[object Map]',
+  Date: '[object Date]'
+};
+
 /**
  * Retrieves the prototype.toString() of a value.
  * If the value is an object, it will cache the result in a WeakMap for future use.
@@ -18,37 +29,37 @@ function getPrototypeString(value: unknown): string {
 
 export function isAnyArrayBuffer(value: unknown): value is ArrayBuffer {
   const type = getPrototypeString(value);
-  return ['[object ArrayBuffer]', '[object SharedArrayBuffer]'].includes(type);
+  return type === TYPES.ArrayBuffer || type === TYPES.SharedArrayBuffer;
 }
 
 export function isUint8Array(value: unknown): value is Uint8Array {
   const type = getPrototypeString(value);
-  return type === '[object Uint8Array]';
+  return type === TYPES.Uint8Array;
 }
 
 export function isBigInt64Array(value: unknown): value is BigInt64Array {
   const type = getPrototypeString(value);
-  return type === '[object BigInt64Array]';
+  return type === TYPES.BigInt64Array;
 }
 
 export function isBigUInt64Array(value: unknown): value is BigUint64Array {
   const type = getPrototypeString(value);
-  return type === '[object BigUint64Array]';
+  return type === TYPES.BigUint64Array;
 }
 
 export function isRegExp(d: unknown): d is RegExp {
   const type = getPrototypeString(d);
-  return type === '[object RegExp]';
+  return type === TYPES.RegExp;
 }
 
 export function isMap(d: unknown): d is Map<unknown, unknown> {
   const type = getPrototypeString(d);
-  return type === '[object Map]';
+  return type === TYPES.Map;
 }
 
 export function isDate(d: unknown): d is Date {
   const type = getPrototypeString(d);
-  return type === '[object Date]';
+  return type === TYPES.Date;
 }
 
 export type InspectFn = (x: unknown, options?: unknown) => string;