### Description

This MR improves serialization (and `calculateSize`) performance by up to 20-40% in my tests using MFlix documents.

| Test Case               | Ops/sec (previous) | Ops/sec (new) | Improvement (%) |
|-------------------------|----------------|-------------------|-----------------|
| `serialize`             | 3,102,374      | 3,752,681         | 20.96%          |
| `calculateSize`         | 23,983,120     | 33,635,923        | 40.25%          |
| `serialize large doc`   | 719,650        | 926,643           | 28.76%          |


- `Object.prototype.toString.call(x)` seems to be quite expensive
- Re-order serialization type checks to move these calls out of hot paths
- Add WeakMap cache for repeated calls (ie. calling isDate(x), isRegExp(x), isUint8Array(x) will only result in 1 `Object.prototype.toString.call(x)`)
- Add global symbol constant so it doesn't have to search the symbol registry for every property

#### What is changing?

It may look like a lot of changes, but most of the changes here are just re-ordering the type checks during serialization. There shouldn't be any functionality changes here and all tests are passing.

##### Is there new documentation needed for these changes?

None

#### What is the motivation for this change?

<!-- If this is a bug, it helps to describe the current behavior and a clear outline of the expected behavior -->
<!-- If this is a feature, it helps to describe the new use case enabled by this change -->

<!--
Contributors!
First of all, thank you so much!!
If you haven't already, it would greatly help the team review this work in a timely manner if you create a JIRA ticket to track this PR.
You can do that here: https://jira.mongodb.org/projects/NODE
-->

### Release Highlight

<!-- RELEASE_HIGHLIGHT_START -->

### Serialization performance improved.

Optimizations have been implemented with respect to BSON serialization across the board, resulting in up to 20% gains in serialization with a sample of [MFlix](https://www.mongodb.com/docs/atlas/sample-data/sample-mflix/) documents. Many thanks to @SeanReece for the contribution.

<!-- RELEASE_HIGHLIGHT_END -->

### Double check the following

- [x] Ran `npm run check:lint` script
- [x] Self-review completed using the [steps outlined here](https://github.com/mongodb/node-mongodb-native/blob/HEAD/CONTRIBUTING.md#reviewer-guidelines)
- [x] PR title follows the [correct format](https://www.conventionalcommits.org/en/v1.0.0/): `type(NODE-xxxx)[!]: description`
  - Example: `feat(NODE-1234)!: rewriting everything in coffeescript`
- [x] Changes are covered by tests
- [x] New TODOs have a related JIRA ticket