@durran 2024-09-05
At the PR's current state the standouts on our benchmarks:

Notable winners:

| Test Case                               | MB/sec (prev) | MB/sec (new) | Improvement (%) |
|-----------------------------------------|---------------|--------------|-----------------|
| timestamp_array_1000._serialize_bson    |         93.16 |       201.89 |          116.7% |
| objectid_array_1000._serialize_bson     |        121.31 |       224.35 |           84.9% |
| decimal128_array_1000._serialize_bson   |        144.92 |       263.05 |           81.5% |
| binary_array_1000._serialize_bson       |        250.30 |       439.65 |           75.6% |

Notable losers:


| Test Case                               | MB/sec (prev) | MB/sec (new) |    Decrease (%) |
|-----------------------------------------|---------------|--------------|-----------------|
| decimal128_array_1000._deserialize_bson |         46.88 |        22.24 |         -110.8% |
| nested_16._serialize_bson               |         37.85 |        29.44 |          -28.6% |
| mixed_large._deserialize_bson           |         80.60 |        73.77 |           -9.2% |


So across the board this impacts large array serialization dramatically - huge gains everywhere. But the deserialization of decimal 128 losses we need to look into as well as some of the other outliers. Nothing notable outside of that from our granular benchmarks.