@SeanReece 2024-09-05
@durran That's great news, very odd about the deserialize regression. I've been struggling getting bson-bench running on my M1 Mac, but I was able to get it to run in docker. I'll do some tests and see if I can reproduce.