@SeanReece 2024-09-16
@durran I was able to reproduce the issue with `decimal128` deserialize. Since the decimal128 constructor calls, `isUint8Array()` there was some extra overhead for that single call, adding a `instanceof` check improves the performance over over main by up to 93% on my machine.

Winners 🎉 
| Name                                           | MB/sec (prev) | MB/sec (new) | Improvement (%) |
|------------------------------------------------|--------------|------------|---------|
| maxkey_array_1000._serialize_bson              | 108.2MB      | 390.3MB    | 260.7%  |
| minkey_array_1000._serialize_bson              | 112.0MB      | 392.7MB    | 250.6%  |
| minkey_array_100._serialize_bson               | 89.3MB       | 275.8MB    | 209.0%  |
| maxkey_array_100._serialize_bson               | 88.4MB       | 260.9MB    | 195.1%  |
| timestamp_array_1000._serialize_bson           | 269.0MB      | 793.2MB    | 194.9%  |
| timestamp_array_100._serialize_bson            | 225.6MB      | 651.2MB    | 188.6%  |
| objectid_array_1000._serialize_bson            | 329.1MB      | 875.9MB    | 166.2%  |
| objectid_array_100._serialize_bson             | 307.9MB      | 739.1MB    | 140.0%  |
| decimal128_array_1000._deserialize_bson        | 130.6MB      | 283.0MB    | 116.8%  |
| decimal128_array_100._deserialize_bson         | 124.3MB      | 258.4MB    | 107.9%  |
| decimal128_array_1000._serialize_bson          | 313.9MB      | 607.2MB    | 93.4%   |
| binary_array_1000._serialize_bson              | 607.9MB      | 1148.4MB   | 88.9%   |
| binary_array_100._serialize_bson               | 578.6MB      | 1088.4MB   | 88.1%   |
| decimal128_array_100._serialize_bson           | 302.2MB      | 552.3MB    | 82.8%   |
| decimal128_array_10._deserialize_bson          | 109.4MB      | 183.1MB    | 67.3%   |
| minkey_array_10._serialize_bson                | 45.7MB       | 74.2MB     | 62.5%   |
| maxkey_array_10._serialize_bson                | 48.0MB       | 76.2MB     | 58.9%   |
| binary_array_10._serialize_bson                | 304.1MB      | 454.3MB    | 49.4%   |
| null_array_1000._serialize_bson                | 337.8MB      | 498.9MB    | 47.7%   |
| timestamp_array_10._serialize_bson             | 108.2MB      | 159.6MB    | 47.4%   |
| null_array_100._serialize_bson                 | 235.1MB      | 312.3MB    | 32.8%   |
| mixed_small._deserialize_bson                  | 155.0MB      | 204.2MB    | 31.8%   |
| code-without-scope_array_100._serialize_bson   | 838.8MB      | 1073.2MB   | 27.9%   |
| objectid_array_10._serialize_bson              | 138.3MB      | 176.4MB    | 27.6%   |
| code-without-scope_array_10._serialize_bson    | 638.8MB      | 812.5MB    | 27.2%   |
| null_singleElementArray._serialize_bson        | 15.5MB       | 19.5MB     | 26.2%   |
| decimal128_array_10._serialize_bson            | 145.2MB      | 181.5MB    | 25.0%   |
| regex_array_10._serialize_bson                 | 222.1MB      | 277.0MB    | 24.7%   |
| code-with-scope_singleFieldDocument._serialize_bson | 135.4MB | 167.3MB   | 23.6%   |
| null_array_10._serialize_bson                  | 33.8MB       | 41.5MB     | 22.9%   |
| regex_array_1000._serialize_bson               | 304.8MB      | 372.2MB    | 22.1%   |
| decimal128_singleFieldDocument._deserialize_bson | 23.2MB     | 28.2MB     | 21.3%   |
| code-without-scope_array_1000._serialize_bson  | 922.7MB      | 1115.4MB   | 20.9%   |
| regex_array_100._serialize_bson                | 297.2MB      | 348.6MB    | 17.3%   |
| code-with-scope_singleElementArray._serialize_bson | 122.6MB   | 143.8MB   | 17.2%   |
| date_array_10._serialize_bson                  | 100.2MB      | 115.9MB    | 15.7%   |
| objectid_singleFieldDocument._serialize_bson   | 25.1MB       | 29.0MB     | 15.4%   |
| binary_small._serialize_bson                   | 1069.7MB     | 1221.1MB   | 14.1%   |
| timestamp_singleFieldDocument._serialize_bson  | 19.6MB       | 22.4MB     | 14.1%   |
| decimal128_singleFieldDocument._serialize_bson | 30.1MB       | 34.0MB     | 13.2%   |
| binary_singleFieldDocument._serialize_bson     | 48.6MB       | 55.0MB     | 13.1%   |
| code-without-scope_singleFieldDocument._serialize_bson | 127.8MB  | 143.5MB   | 12.3%   |
| code-with-scope_array_100._serialize_bson      | 535.7MB      | 599.0MB    | 11.8%   |
| null_singleFieldDocument._serialize_bson       | 12.5MB       | 13.9MB     | 11.0%   |
| code-with-scope_array_1000._serialize_bson     | 518.6MB      | 575.5MB    | 11.0%   |
| long_array_1000._deserialize_bson              | 327.1MB      | 362.1MB    | 10.7%   |
| maxkey_singleFieldDocument._serialize_bson     | 8.5MB        | 9.4MB      | 10.6%   |

Losers: None